+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/model/upsert/output/OutputUpsertData.kt
@@ -51,5 +51,7 @@ data class OutputUpsertData(
     val addressUpdatedAt: Instant?,
     val scriptVariants: List<BusinessPartnerScriptVariantDto>,
     val legalEntityGoldenRecordRelations: List<LegalEntityGoldenRecordRelation>,
-    val addressGoldenRecordRelations: List<AddressGoldenRecordRelation>
+    val addressGoldenRecordRelations: List<AddressGoldenRecordRelation>,
+    val ownershipUltimate: Boolean? = null,
+    val ultimateOwnerBpnl: String? = null
 )