+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/GoldenRecordUpdateServices.kt
@@ -247,6 +247,8 @@ class GoldenRecordUpdateChunkService(
             scriptVariants = scriptVariants.map { businessPartnerMappings.toScriptVariantDto(it) },
             legalEntityGoldenRecordRelations = legalEntityGoldenRecordRelations.map { it.toUpsertData() },
             addressGoldenRecordRelations = addressGoldenRecordRelations.map { it.toUpsertData() },
+            ownershipUltimate = ownershipUltimate,
+            ultimateOwnerBpnl = ultimateOwnerBpnl
         )
     }
 
@@ -339,6 +341,8 @@ class GoldenRecordUpdateChunkService(
         businessPartner.addressUpdatedAt = legalEntity.legalAddress.updatedAt
         businessPartner.legalEntityConfidence?.let { update(it,  header.confidenceCriteria) }
         businessPartner.legalEntityGoldenRecordRelations.addAll(legalEntity.header.relations.map(::toEntity))
+        businessPartner.ownershipUltimate = header.ownershipUltimate
+        businessPartner.ultimateOwnerBpnl = header.ultimateOwnerBpnl
 
         val variantByCode = businessPartner.scriptVariants.associateBy { it.scriptCode }