+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/OutputUpsertMappings.kt
@@ -50,6 +50,8 @@ class OutputUpsertMappings(
                 bpnL = legalEntityBpn,
                 bpnS = siteBpn,
                 bpnA = addressBpn,
+                ownershipUltimate = ownershipUltimate,
+                ultimateOwnerBpnl = ultimateOwnerBpnl,
                 postalAddress = PostalAddressDb(addressType, physicalPostalAddress.toEntity(), alternativePostalAddress?.toEntity()),
                 legalEntityConfidence = legalEntityConfidence.toEntity(),
                 siteConfidence = siteConfidence?.toEntity(),