+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/util/BusinessPartnerCopyUtil.kt
@@ -43,6 +43,8 @@ class BusinessPartnerCopyUtil {
             bpnL = fromPartner.bpnL
             bpnS = fromPartner.bpnS
             bpnA = fromPartner.bpnA
+            ownershipUltimate = fromPartner.ownershipUltimate
+            ultimateOwnerBpnl = fromPartner.ultimateOwnerBpnl
             legalEntityConfidence = fromPartner.legalEntityConfidence
             siteConfidence = fromPartner.siteConfidence
             addressConfidence = fromPartner.addressConfidence