+++ bpdm-gate/src/main/resources/db/migration/V7_5_0_2__add_ultimate_owner_columns.sql
@@ -1,0 +1,3 @@
+ALTER TABLE business_partners
+ADD COLUMN ownership_ultimate BOOLEAN,
+ADD COLUMN ultimate_owner_bpnl VARCHAR(20);