+++ bpdm-system-tester/src/main/kotlin/org/eclipse/tractusx/bpdm/test/system/utils/GateOutputFactory.kt
@@ -65,6 +65,8 @@ class GateOutputFactory(
                     shortName = shortName,
                     legalForm = legalForm,
                     confidenceCriteria = dummyConfidence,
+                    ownershipUltimate = false,
+                    ultimateOwnerBpnl = null,
                     states = states
                 )
             },